refactor: active와 pending으로 분리
